Disable config file watcher if a hot-reload capability is set (#128)
Corresponds to https://reviews.llvm.org/D94222

This is the minimal change. I have a feeling the dispose/re-watch logic could be made cleaner.
But this client-side feature can just be removed in future, so I'm not sure it's worth the risk of trying to polish it.

I've tested this locally with clangd with/without the capability.