Inactive regions support via dedicated protocol (#193)
If the server supports the `textDocument/inactiveRegions` protocol extension (added in clangd 17), use it to apply a style to the inactive regions that combines with the foreground colors from the client-side (textmate) tokens rather than overwriting them.

The default inactive region style is reduced opacity. The opacity level can be customized with `clangd.inactiveRegions.opacity`.

An alternative inactive region style of a background highlight can be enabled with `clangd.inactiveRegions.useBackgroundHighlight=true`. The highlight color can be customized with `clangd.inactiveRegions.background` in workbench.colorCustomizations.